pyproject.toml
[project]
name = "texa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markdown>=3.7",
    "pillow>=10.4.0",
    "pyside6>=6.7.2",
    "texify>=0.1.10",
    "pymdown-extensions>=10.11",
]

[tool.pyright]
typeCheckingMode = "strict"
reportUnknownMemberType = false
reportAny = false
reportImplicitOverride = false
reportMissingTypeStubs = false

[tool.ruff.lint]
extend-select = ["I", "N", "PERF", "UP"]
